refactor(ComboBox): moved ComboBoxRequestStatus to constants
