+++ packages/retail-ui/components/CustomComboBox/constants.ts
@@ -1,0 +1,6 @@
+export enum ComboBoxRequestStatus {
+  Pending,
+  Success,
+  Failed,
+  Unknown
+}
