+++ packages/retail-ui/components/CustomComboBox/reducer/default.tsx
@@ -4,7 +4,6 @@ import isEqual from 'lodash.isequal';
 
 import Menu from '../../Menu/Menu';
 import CustomComboBox, {
-  ComboBoxRequestStatus,
   CustomComboBoxProps,
   CustomComboBoxState,
   DefaultState
@@ -14,6 +13,7 @@ import { Nullable } from '../../../typings/utility-types';
 import warning from 'warning';
 import ComboBoxView from '../ComboBoxView';
 import reactGetTextContent from '../../../lib/reactGetTextContent/reactGetTextContent';
+import { ComboBoxRequestStatus } from '../constants';
 
 interface BaseAction {
   type: string;
